Provo a pilotare le dimensioni dello slide show

+++ Digiphoto.Lumen.UI/Pubblico/SlideShow.cs
@@ -15,12 +15,12 @@ namespace Digiphoto.Lumen.UI.Pubblico {
 
 	public class SlideShow {
 
-		public int righe {
+		public short righe {
 			get;
 			set;
 		}
 
-		public int colonne {
+		public short colonne {
 			get;
 			set;
 		}
